Update and rename RunningSumArray1.java to RunningSum.java
